refactor(metal): move the dense decode encode path into its own module
attn.rs is 9317 lines, which is why nobody reads the decode encode loop
end to end, and GitHub issue #149 says that loop's HOST cost, not its
kernels, is the whole remaining Metal decode gap. Pull
launch_decode_dense_stack and the types only it uses into
decode_dense.rs so the thing being measured fits on a screen.

Pure move: the function body is byte-identical, and the only edits are
pub(crate) on the helpers it now reaches across the module boundary, a
pub use in attn.rs so no caller changes, and pointing
every_encode_in_the_decode_stack_declares_its_hazards at the file the
stack now lives in.
