+++ crates/ferrox-metal/src/decode_dense.rs
@@ -1,0 +1,576 @@
+//! The dense B=1 decode stack: every layer of a dense model in one
+//! command buffer, hidden state resident on the GPU across layers.
+//!
+//! Split out of `attn.rs` along the seam GitHub issue #149 touches: the
+//! host-side cost of encoding this stack is the whole remaining Metal
+//! decode gap, so the encode loop needs to live somewhere it can be read
+//! and changed in full. Per-token GPU/host accounting is in
+//! `crate::gpu::gpu_timing_note` and `crate::mem_ranges`.
+
+use crate::attn::{
+    assert_freq_factors_len, borrow_decode_scratch, copy_f32_into, encode_attn_extras,
+    encode_gqa_with_kv, encode_kv_store_append, encode_rope, KvPlane, LayerRope, MetalKvBuffers,
+    MetalRope, ScratchCaps,
+};
+use crate::elem::{
+    encode_add_rms_norm, encode_argmax, encode_gelu_mul, encode_rms_norm, encode_silu_mul,
+    encode_vec_add,
+};
+use crate::embd::{encode_get_rows, EmbdKind};
+use crate::gpu::{
+    compute_encoder_concurrent, encode_matvec, resident_f32_buffer, resident_weight_buffer,
+    shared_metal, MatvecLaunch, MetalError,
+};
+use crate::mem_ranges::MemRanges;
+use objc2::runtime::ProtocolObject;
+use objc2_metal::{MTLBuffer, MTLCommandBuffer, MTLCommandEncoder, MTLCommandQueue};
+
+/// Optional attention epilogue ops applied between the QKV matvecs and
+/// RoPE, in CPU-path order: bias add (Qwen2-family `qkv_bias`), then
+/// QK-RMSNorm — per-head (Qwen3 / Gemma-3, `weight.len() == head_dim`)
+/// or whole-vector (OLMoE, `weight.len() == n_heads|n_kv_heads * head_dim`).
+/// `attn_logit_softcap` is applied inside GQA after score scaling
+/// (Gemma-2); when set, FA-vec is skipped in favour of the legacy kernel
+/// unless the FA-vec softcap path is enabled.
+#[derive(Default)]
+pub struct AttnExtras<'a> {
+    pub q_bias: Option<&'a [f32]>,
+    pub k_bias: Option<&'a [f32]>,
+    pub v_bias: Option<&'a [f32]>,
+    pub q_norm: Option<&'a [f32]>,
+    pub k_norm: Option<&'a [f32]>,
+    pub attn_logit_softcap: Option<f32>,
+}
+
+impl AttnExtras<'_> {
+    pub fn is_empty(&self) -> bool {
+        self.q_bias.is_none()
+            && self.k_bias.is_none()
+            && self.v_bias.is_none()
+            && self.q_norm.is_none()
+            && self.k_norm.is_none()
+            && self.attn_logit_softcap.is_none()
+    }
+}
+
+/// Per-layer launches + norms for [`launch_decode_dense_stack`].
+pub struct DenseLayerMetal<'a> {
+    pub attn_norm_w: &'a [f32],
+    pub ffn_norm_w: &'a [f32],
+    pub q: MatvecLaunch<'a>,
+    pub k: MatvecLaunch<'a>,
+    pub v: MatvecLaunch<'a>,
+    pub o: MatvecLaunch<'a>,
+    pub gate: MatvecLaunch<'a>,
+    pub up: MatvecLaunch<'a>,
+    pub down: MatvecLaunch<'a>,
+    pub extras: AttnExtras<'a>,
+    /// This layer's RoPE base AND divisors. Both halves, always: a
+    /// Gemma-3 SWA layer differs from its full-attention neighbours in
+    /// both (`rope_theta_swa`, and no linear scale folded into the
+    /// divisors). See [`LayerRope`].
+    pub rope: LayerRope<'a>,
+    /// Sliding-window size for this layer (`None` = full causal).
+    pub window: Option<usize>,
+    /// Gemma post-attention / post-FFN sandwich norms, applied to the
+    /// block output *before* the residual add.
+    pub post_attn_norm: Option<&'a [f32]>,
+    pub post_ffn_norm: Option<&'a [f32]>,
+}
+
+/// Optional on-GPU embedding gather at the start of
+/// [`launch_decode_dense_stack`] (skips host `dequant_row` + upload).
+pub struct EmbdGatherMetal<'a> {
+    pub kind: EmbdKind,
+    pub weights: &'a [u8],
+    pub rows: usize,
+    pub row_bytes: usize,
+    pub n_cols: usize,
+    pub token_id: usize,
+}
+
+/// All dense layers in **one** command buffer (one wait). Hidden stays on
+/// GPU across layers — Crane-style residency for B=1 decode.
+/// When `embd` is `Some`, gathers that token row into scratch `h` on-GPU
+/// instead of copying a host-provided `hidden` slice.
+/// When `final_norm_w` + `output` are provided, also runs final RMSNorm +
+/// lm_head on-GPU. With `argmax_only`, runs argmax and returns a
+/// **1-element** `vec![token_id as f32]`; otherwise downloads vocab logits.
+///
+/// Chunked multi-CB early-commit (llama `n_main` style) was tried on Host B
+/// and regressed decode tok/s — see `…_multicb*` receipts; kept single CB.
+#[allow(clippy::too_many_arguments)]
+pub fn launch_decode_dense_stack(
+    hidden: &[f32],
+    layers: &[DenseLayerMetal<'_>],
+    kvs: &mut [MetalKvBuffers],
+    n_heads: usize,
+    rope_layout: MetalRope,
+    pos: usize,
+    rms_eps: f32,
+    final_norm_w: Option<&[f32]>,
+    output: Option<&MatvecLaunch<'_>>,
+    argmax_only: bool,
+    embd: Option<&EmbdGatherMetal<'_>>,
+    gelu_ffn: bool,
+) -> Result<Vec<f32>, MetalError> {
+    assert_eq!(layers.len(), kvs.len());
+    assert!(!layers.is_empty());
+    let hidden_dim = match embd {
+        Some(e) => e.n_cols,
+        None => hidden.len(),
+    };
+    assert!(hidden_dim > 0);
+    let head_dim = kvs[0].head_dim;
+    let n_kv_heads = kvs[0].n_kv_heads;
+    for kv in kvs.iter() {
+        assert_eq!(kv.head_dim, head_dim);
+        assert_eq!(kv.n_kv_heads, n_kv_heads);
+        assert_eq!(pos, kv.seq_len);
+        if kv.seq_len >= kv.capacity {
+            return Err(MetalError::CommandFailed);
+        }
+    }
+    for layer in layers.iter() {
+        assert_freq_factors_len(layer.rope.freq_factors, rope_layout, head_dim);
+    }
+
+    let max_q = layers.iter().map(|l| l.q.rows).max().unwrap();
+    let max_kv = layers.iter().map(|l| l.k.rows).max().unwrap();
+    let max_gate = layers.iter().map(|l| l.gate.rows).max().unwrap();
+    let attn_elems = n_heads * head_dim;
+    let logits_rows = output.map(|o| o.rows);
+
+    let shared = shared_metal()?;
+    let device = &shared.device;
+    let queue = &shared.queue;
+
+    let scratch_guard = borrow_decode_scratch(
+        device,
+        ScratchCaps {
+            hidden: hidden_dim,
+            max_q,
+            max_kv,
+            attn: attn_elems,
+            max_gate,
+            logits: logits_rows.unwrap_or(0),
+        },
+    )?;
+    let scratch = scratch_guard.as_ref().expect("scratch just ensured");
+    let h_buf = &scratch.h;
+    if let Some(e) = embd {
+        assert_eq!(e.n_cols, hidden_dim);
+        assert!(e.token_id < e.rows);
+        assert_eq!(e.weights.len(), e.rows * e.row_bytes);
+        // Gather runs in the same CB below (after encoder create).
+    } else {
+        assert_eq!(hidden.len(), hidden_dim);
+        copy_f32_into(h_buf, hidden);
+    }
+    let x_buf = &scratch.x;
+    let x2_buf = &scratch.x2;
+    let q_buf = &scratch.q;
+    let k_buf = &scratch.k;
+    let v_buf = &scratch.v;
+    let attn_buf = &scratch.attn;
+    let o_buf = &scratch.o;
+    let gate_buf = &scratch.gate;
+    let up_buf = &scratch.up;
+    let act_buf = &scratch.act;
+    let down_buf = &scratch.down;
+    let logits_buf = scratch.logits.as_ref();
+    let argmax_idx_buf = &scratch.argmax_idx;
+
+    // One resident buffer PER LAYER. `resident_f32_buffer` keys its
+    // cache on (pointer, len), so the at-most-two distinct divisor sets
+    // an alternating-SWA model has are uploaded once each and every
+    // layer past the first two is a cache hit -- no per-layer upload,
+    // and no table of "which set does layer i use" for a call site to
+    // get out of step with.
+    let ff_resident = layers
+        .iter()
+        .map(|l| match l.rope.freq_factors {
+            Some(ff) => resident_f32_buffer(device, ff).map(Some),
+            None => Ok(None),
+        })
+        .collect::<Result<Vec<_>, MetalError>>()?;
+
+    let cmd_buf = queue.commandBuffer().ok_or(MetalError::CommandFailed)?;
+    // Gemma-style post-norms ("sandwich"): these layers take the EAGER
+    // residual path below, which is what makes concurrent encode safe here.
+    let sandwich = layers
+        .iter()
+        .any(|l| l.post_attn_norm.is_some() || l.post_ffn_norm.is_some());
+    // Every model encodes concurrently: gate∥up and Q∥K∥V overlap, and the
+    // hazard tracker emits a barrier only where a dispatch reads or
+    // overwrites something still in flight, narrowed to those resources.
+    //
+    // Sandwich models (Gemma post-norms) used to be forced onto the serial
+    // encoder, because concurrent dispatch with in-place RMSNorm and
+    // DEFERRED residuals diverged from CPU on Gemma-2 B=1 decode. That fix
+    // landed two changes at once -- serial encode AND eager residuals -- and
+    // the eager residuals are the half that mattered: with them, every op in
+    // this function declares its own reads and writes (`encode_gqa_with_kv`
+    // self-tracks, including the f16 dequant scratch no caller can name), so
+    // concurrency is safe by construction rather than by scheduling luck.
+    //
+    // Measured on an M2 Pro, interleaved, GPU-clock: Gemma-2-2B Q4_K_M decode
+    // 13.23 -> 12.20 ms/token, and greedy output stays byte-identical to the
+    // serial encoder across Gemma-2 and Gemma-3 on every prompt tried.
+    let (encoder, mut mrs) = (compute_encoder_concurrent(&cmd_buf)?, MemRanges::new());
+
+    let embd_resident = if let Some(e) = embd {
+        let w = resident_weight_buffer(device, e.weights)?;
+        mrs.begin_op(&encoder, &[], &[h_buf]);
+        encode_get_rows(
+            &encoder,
+            device,
+            e.kind,
+            &w,
+            h_buf,
+            e.row_bytes as u32,
+            e.n_cols as u32,
+            e.token_id as u32,
+        )?;
+        mrs.end_op(&[], &[h_buf]);
+        Some(w)
+    } else {
+        None
+    };
+    let _embd_resident = embd_resident;
+
+    // Gemma sandwich (post_attn / post_ffn) must apply residuals eagerly —
+    // same shape as the working prefill stack / CPU path. Deferred
+    // `h += down` fused into the next layer's attn_norm matches SmolLM2
+    // (no post-norms) but diverges for Gemma-2 Metal greedy (BOS loops /
+    // `*` spam) even when GQA unit tests pass.
+    // `sandwich` was computed above (also selects serial encoder).
+
+    for (layer_idx, (layer, kv)) in layers.iter().zip(kvs.iter_mut()).enumerate() {
+        assert_eq!(layer.attn_norm_w.len(), hidden_dim);
+        assert_eq!(layer.ffn_norm_w.len(), hidden_dim);
+        assert_eq!(layer.o.rows, hidden_dim);
+        assert_eq!(layer.down.rows, hidden_dim);
+        assert_eq!(layer.gate.rows, layer.up.rows);
+
+        let attn_nw = resident_f32_buffer(device, layer.attn_norm_w)?;
+        let ffn_nw = resident_f32_buffer(device, layer.ffn_norm_w)?;
+        let q_w = resident_weight_buffer(device, layer.q.weights)?;
+        let k_w = resident_weight_buffer(device, layer.k.weights)?;
+        let v_w = resident_weight_buffer(device, layer.v.weights)?;
+        let o_w = resident_weight_buffer(device, layer.o.weights)?;
+        let gate_w = resident_weight_buffer(device, layer.gate.weights)?;
+        let up_w = resident_weight_buffer(device, layer.up.weights)?;
+        let down_w = resident_weight_buffer(device, layer.down.weights)?;
+        let kv_k = kv.k.as_ref();
+        let kv_v = kv.v.as_ref();
+
+        // Pre-LN: layer 0 norms raw hidden; later layers either fuse the
+        // previous FFN residual into attn_norm (non-sandwich) or just
+        // RMSNorm (sandwich already applied `h += down` eagerly).
+        if layer_idx == 0 || sandwich {
+            mrs.begin_op(&encoder, &[h_buf], &[x_buf]);
+            encode_rms_norm(
+                &encoder,
+                device,
+                h_buf,
+                &attn_nw.buffer,
+                x_buf,
+                hidden_dim as u32,
+                rms_eps,
+            )?;
+            mrs.end_op(&[h_buf], &[x_buf]);
+        } else {
+            mrs.begin_op(&encoder, &[h_buf, down_buf], &[h_buf, x_buf]);
+            encode_add_rms_norm(
+                &encoder,
+                device,
+                h_buf,
+                down_buf,
+                &attn_nw.buffer,
+                x_buf,
+                hidden_dim as u32,
+                rms_eps,
+            )?;
+            mrs.end_op(&[h_buf, down_buf], &[h_buf, x_buf]);
+        }
+        // Q∥K∥V
+        mrs.begin_op(&encoder, &[x_buf], &[q_buf, k_buf, v_buf]);
+        encode_matvec(&encoder, device, &layer.q, &q_w, x_buf, q_buf)?;
+        encode_matvec(&encoder, device, &layer.k, &k_w, x_buf, k_buf)?;
+        encode_matvec(&encoder, device, &layer.v, &v_w, x_buf, v_buf)?;
+        mrs.end_op(&[x_buf], &[q_buf, k_buf, v_buf]);
+
+        mrs.begin_op(&encoder, &[q_buf, k_buf, v_buf], &[q_buf, k_buf, v_buf]);
+        encode_attn_extras(
+            &encoder,
+            device,
+            &layer.extras,
+            q_buf,
+            k_buf,
+            v_buf,
+            layer.q.rows,
+            layer.k.rows,
+            layer.v.rows,
+            n_heads,
+            n_kv_heads,
+            head_dim,
+            rms_eps,
+        )?;
+        mrs.end_op(&[q_buf, k_buf, v_buf], &[q_buf, k_buf, v_buf]);
+
+        let layer_theta = layer.rope.theta;
+        let ff_buf = ff_resident[layer_idx].as_ref().map(|b| b.buffer.as_ref());
+        mrs.begin_op(&encoder, &[q_buf, k_buf], &[q_buf, k_buf]);
+        encode_rope(
+            &encoder,
+            device,
+            rope_layout,
+            q_buf,
+            n_heads as u32,
+            head_dim as u32,
+            layer_theta,
+            pos as u32,
+            ff_buf,
+        )?;
+        encode_rope(
+            &encoder,
+            device,
+            rope_layout,
+            k_buf,
+            n_kv_heads as u32,
+            head_dim as u32,
+            layer_theta,
+            pos as u32,
+            ff_buf,
+        )?;
+        mrs.end_op(&[q_buf, k_buf], &[q_buf, k_buf]);
+
+        let token_elems = (n_kv_heads * head_dim) as u32;
+        let offset = (pos * n_kv_heads * head_dim) as u32;
+        mrs.begin_op(&encoder, &[k_buf, v_buf], &[kv_k, kv_v]);
+        encode_kv_store_append(&encoder, device, k_buf, kv, KvPlane::K, offset, token_elems)?;
+        encode_kv_store_append(&encoder, device, v_buf, kv, KvPlane::V, offset, token_elems)?;
+        mrs.end_op(&[k_buf, v_buf], &[kv_k, kv_v]);
+
+        let new_seq = (pos + 1) as u32;
+        // Sliding window: only the last `window` positions (incl. current)
+        // are visible, matching `causal_gqa_attention_windowed`.
+        let kv_start = match layer.window {
+            Some(w) => (pos + 1).saturating_sub(w) as u32,
+            None => 0,
+        };
+        // Self-tracking: with a quantized KV cache this also writes a shared
+        // f16 dequant scratch that no caller can name.
+        encode_gqa_with_kv(
+            &encoder,
+            &mut mrs,
+            device,
+            q_buf,
+            kv,
+            attn_buf,
+            n_heads as u32,
+            n_kv_heads as u32,
+            head_dim as u32,
+            new_seq,
+            kv_start,
+            layer.extras.attn_logit_softcap,
+        )?;
+
+        mrs.begin_op(&encoder, &[attn_buf], &[o_buf]);
+        encode_matvec(&encoder, device, &layer.o, &o_w, attn_buf, o_buf)?;
+        mrs.end_op(&[attn_buf], &[o_buf]);
+
+        // Gemma sandwich norm: normalize the attn block output *before*
+        // the residual add (in-place: each thread reads x[i] only after
+        // the barriered reduction, so out == x is safe).
+        if let Some(post) = layer.post_attn_norm {
+            assert_eq!(post.len(), hidden_dim);
+            let pw = resident_f32_buffer(device, post)?;
+            mrs.begin_op(&encoder, &[o_buf], &[o_buf]);
+            encode_rms_norm(
+                &encoder,
+                device,
+                o_buf,
+                &pw.buffer,
+                o_buf,
+                hidden_dim as u32,
+                rms_eps,
+            )?;
+            mrs.end_op(&[o_buf], &[o_buf]);
+        }
+        // Attn residual + ffn_norm in one dispatch, for every model.
+        //
+        // Sandwich layers used to split this into `vec_add` then `rms_norm`,
+        // which is the same arithmetic in two dispatches: `post_attn_norm`
+        // has already been applied to `o_buf` in place above, so both paths
+        // compute `h += o` then `x2 = rms_norm(h)`. The split was a leftover
+        // from the serial-encoder era -- `encode_add_rms_norm` writes `h`
+        // itself, so the residual is just as eager as the two-dispatch form.
+        mrs.begin_op(&encoder, &[h_buf, o_buf], &[h_buf, x2_buf]);
+        encode_add_rms_norm(
+            &encoder,
+            device,
+            h_buf,
+            o_buf,
+            &ffn_nw.buffer,
+            x2_buf,
+            hidden_dim as u32,
+            rms_eps,
+        )?;
+        mrs.end_op(&[h_buf, o_buf], &[h_buf, x2_buf]);
+        // gate ∥ up (llama concurrent)
+        mrs.begin_op(&encoder, &[x2_buf], &[gate_buf, up_buf]);
+        encode_matvec(&encoder, device, &layer.gate, &gate_w, x2_buf, gate_buf)?;
+        encode_matvec(&encoder, device, &layer.up, &up_w, x2_buf, up_buf)?;
+        mrs.end_op(&[x2_buf], &[gate_buf, up_buf]);
+
+        mrs.begin_op(&encoder, &[gate_buf, up_buf], &[act_buf]);
+        if gelu_ffn {
+            encode_gelu_mul(
+                &encoder,
+                device,
+                gate_buf,
+                up_buf,
+                act_buf,
+                layer.gate.rows as u32,
+            )?;
+        } else {
+            encode_silu_mul(
+                &encoder,
+                device,
+                gate_buf,
+                up_buf,
+                act_buf,
+                layer.gate.rows as u32,
+            )?;
+        }
+        mrs.end_op(&[gate_buf, up_buf], &[act_buf]);
+
+        mrs.begin_op(&encoder, &[act_buf], &[down_buf]);
+        encode_matvec(&encoder, device, &layer.down, &down_w, act_buf, down_buf)?;
+        mrs.end_op(&[act_buf], &[down_buf]);
+
+        if let Some(post) = layer.post_ffn_norm {
+            assert_eq!(post.len(), hidden_dim);
+            let pw = resident_f32_buffer(device, post)?;
+            mrs.begin_op(&encoder, &[down_buf], &[down_buf]);
+            encode_rms_norm(
+                &encoder,
+                device,
+                down_buf,
+                &pw.buffer,
+                down_buf,
+                hidden_dim as u32,
+                rms_eps,
+            )?;
+            mrs.end_op(&[down_buf], &[down_buf]);
+        }
+        if sandwich {
+            // Eager FFN residual — next layer attn_norm is plain RMSNorm.
+            mrs.begin_op(&encoder, &[h_buf, down_buf], &[h_buf]);
+            encode_vec_add(&encoder, device, h_buf, down_buf, hidden_dim as u32)?;
+            mrs.end_op(&[h_buf, down_buf], &[h_buf]);
+        }
+        // Non-sandwich: defer `h += down` until the next layer's attn_norm
+        // (or final_norm) so it fuses with that RMSNorm. `down_buf` stays in
+        // the tracker's dst set, so the next layer's fused norm barriers
+        // against it exactly once. Last layer handled below.
+    }
+
+    // Final norm / lm_head. Sandwich already applied every FFN residual;
+    // non-sandwich still has a deferred last-layer `down` to fold in.
+    let (download_n, norm_resident) = if let Some(fnw) = final_norm_w {
+        assert_eq!(fnw.len(), hidden_dim);
+        let fn_buf = resident_f32_buffer(device, fnw)?;
+        if sandwich {
+            mrs.begin_op(&encoder, &[h_buf], &[x_buf]);
+            encode_rms_norm(
+                &encoder,
+                device,
+                h_buf,
+                &fn_buf.buffer,
+                x_buf,
+                hidden_dim as u32,
+                rms_eps,
+            )?;
+            mrs.end_op(&[h_buf], &[x_buf]);
+        } else {
+            mrs.begin_op(&encoder, &[h_buf, down_buf], &[h_buf, x_buf]);
+            encode_add_rms_norm(
+                &encoder,
+                device,
+                h_buf,
+                down_buf,
+                &fn_buf.buffer,
+                x_buf,
+                hidden_dim as u32,
+                rms_eps,
+            )?;
+            mrs.end_op(&[h_buf, down_buf], &[h_buf, x_buf]);
+        }
+        if let (Some(out_l), Some(logits)) = (output, logits_buf) {
+            assert_eq!(out_l.rows, logits_rows.unwrap());
+            // RAW: lm_head reads `x_buf` written by the norm above. Without
+            // ordering here Metal may overlap the matvec with the norm on
+            // small hiddens (SmolLM2 h=576) and produce garbage logits /
+            // greedy tokens while host lm_head after wait looks fine — the
+            // tracker sees `x_buf` as src-after-dst and barriers.
+            mrs.begin_op(&encoder, &[x_buf], &[logits.as_ref()]);
+            let out_w = resident_weight_buffer(device, out_l.weights)?;
+            encode_matvec(&encoder, device, out_l, &out_w, x_buf, logits)?;
+            mrs.end_op(&[x_buf], &[logits.as_ref()]);
+            if argmax_only {
+                mrs.begin_op(&encoder, &[logits.as_ref()], &[argmax_idx_buf]);
+                encode_argmax(&encoder, device, logits, argmax_idx_buf, out_l.rows as u32)?;
+                mrs.end_op(&[logits.as_ref()], &[argmax_idx_buf]);
+                (1, false)
+            } else {
+                (out_l.rows, false)
+            }
+        } else {
+            // final_norm ran but no lm_head — download normalized hidden
+            // and mark x_buf resident for the next apply_gpu.
+            (hidden_dim, true)
+        }
+    } else if sandwich {
+        (hidden_dim, false)
+    } else {
+        // No final_norm: still apply the deferred last-layer FFN residual.
+        mrs.begin_op(&encoder, &[h_buf, down_buf], &[h_buf]);
+        encode_vec_add(&encoder, device, h_buf, down_buf, hidden_dim as u32)?;
+        mrs.end_op(&[h_buf, down_buf], &[h_buf]);
+        (hidden_dim, false)
+    };
+
+    encoder.endEncoding();
+    cmd_buf.commit();
+    cmd_buf.waitUntilCompleted();
+    crate::gpu::gpu_timing_note(&cmd_buf, "dense-decode/tok", 32);
+
+    for kv in kvs.iter_mut() {
+        kv.seq_len = pos + 1;
+    }
+
+    // If final_norm ran but no lm_head, mark normalized hidden (x_buf)
+    // as resident so the next apply_gpu can skip re-upload.
+    if norm_resident {
+        crate::gpu::set_resident_activation(x_buf, hidden_dim);
+    }
+
+    if argmax_only && download_n == 1 && output.is_some() {
+        let ptr = argmax_idx_buf.contents();
+        let idx = unsafe { *(ptr.as_ptr() as *const u32) as usize };
+        return Ok(vec![idx as f32]);
+    }
+
+    let src: &ProtocolObject<dyn MTLBuffer> = if norm_resident {
+        x_buf
+    } else if download_n == hidden_dim {
+        h_buf
+    } else {
+        logits_buf.expect("logits buffer when downloading logits")
+    };
+    let out_ptr = src.contents();
+    Ok(unsafe { std::slice::from_raw_parts(out_ptr.as_ptr() as *const f32, download_n).to_vec() })
+}
